Remade modals, now using states when possible. Remade themes. Move LibraryHeader in a separate file. Merge LibraryAll and LibraryNotInShelf in LibraryBase. Fix BookInfoModal for books without a cover.

+++ frontend/src/components/library/modals/BookDeleteModal.js
@@ -1,30 +1,34 @@
-import {useEffect, useRef} from "react";
+import {useEffect, useRef, useState} from "react";
 import {deleteBook} from "../../Api";
 
 function BookDeleteModal(props) {
-    const deleteModal = useRef();
-    const deleteTitle = useRef();
-    const deleteConfirm = useRef();
+    const [id, setId] = useState(null);
+    const [title, setTitle] = useState("...");
+    const modal = useRef();
 
     useEffect(() => {
-        deleteModal.current.addEventListener("show.bs.modal", (e) => {
-            const id = e.relatedTarget.getAttribute("data-bs-id");
-            deleteTitle.current.textContent = e.relatedTarget.getAttribute("data-bs-title");
-            deleteConfirm.current.onclick = () => {
-                deleteBook(id).then(
-                    res => {
-                        console.log("Book", id, "Deleted successfully!");
-                        props.refresh();
-                    },
-                    err => console.error(err)
-                );
-            };
+        modal.current.addEventListener("show.bs.modal", (e) => {
+            setId(e.relatedTarget.getAttribute("data-bs-id"));
+            setTitle(e.relatedTarget.getAttribute("data-bs-title"));
         });
     }, []);
 
+    function confirm() {
+        if (id === null) {
+            return;
+        }
+        deleteBook(id).then(
+            res => {
+                console.log("Book", id, "Deleted successfully!");
+                props.refresh();
+            },
+            err => console.error(err)
+        );
+    }
+
     return (
         <div className={"modal fade"} id={"delete-modal"} tabIndex={-1} aria-labelledby={"delete-modal-label"}
-             aria-hidden={true} ref={deleteModal}>
+             aria-hidden={true} ref={modal}>
             <div className={"modal-dialog"}>
                 <div className={"modal-content"}>
                     <div className={"modal-header"}>
@@ -32,16 +36,15 @@ function BookDeleteModal(props) {
                         <button type={"button"} className={"btn-close"} data-bs-dismiss={"modal"} aria-label={"Close"}/>
                     </div>
                     <div className={"modal-body"}>
-                        <h6 ref={deleteTitle}>...</h6>
+                        <h6>{title}</h6>
                         <p className={"mb-0"}>Are you sure you want to delete this book?</p>
                         <p>This process cannot be undone.</p>
                     </div>
                     <div className={"modal-footer"}>
-                        <button type={"button"} className={"btn btn-secondary"} data-bs-dismiss="modal">
+                        <button type={"button"} className={"btn btn-danger"} data-bs-dismiss="modal">
                             Close
                         </button>
-                        <button type={"button"} className={"btn btn-primary"} data-bs-dismiss="modal"
-                                ref={deleteConfirm}>
+                        <button type={"button"} className={"btn btn-primary"} data-bs-dismiss="modal" onClick={confirm}>
                             Confirm
                         </button>
                     </div>
